requirements.txt
https://download.pytorch.org/whl/cpu/torch-2.0.0%2Bcpu-cp39-cp39-linux_x86_64.whl

faiss-cpu==1.7.4
fastapi==0.95.2
numpy==1.24.3
openai==0.27.7
sentence-transformers==2.2.2
# uvicorn==0.22.0
# websockets==11.0.3